# perf: use context isolation for inidividual journeys