--- a/__tests__/plugins/performance.test.ts
+++ b/__tests__/plugins/performance.test.ts
@@ -59,6 +59,6 @@ describe('performance', () => {
       'JSHeapTotalSize',
     ]);
     await performance.stop();
-    await Gatherer.dispose(driver);
+    await Gatherer.stop();
   });
 });